requirements.txt
pandas
numpy
seaborn
re
transformers
sentence_transformers
langchain
faiss-cpu
tqdm
pickle
langchain-hugggingface